Code_Backup-LSR/discord-bot/requirements.txt
# File: discord-bot/requirements.txt
discord.py
pymysql
